+++ NEWS.md
@@ -1,3 +1,15 @@
 # Version 0.1 (2024/03/28)
 
 * First public release.
+
+
+# Version 0.2 (2024/04/09)
+
+* Friendlier interface for specifying prior distributions.
+
+* Prior distributions now summarised in `summary()` output.
+
+* New function `msmhist()` for illustrating multi-state data over time.
+
+* New function `standardise_to()` for computing outputs standardised
+  over the distribution of a given population.